% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_utils.R
\name{kable_fun}
\alias{kable_fun}
\title{Function to print PDF table}
\usage{
kable_fun(
  dat,
  footnote_arg = list(general_title = "gg: Giorni Stagionatura", general = "")
)
}
\arguments{
\item{dat}{data.frame dati}

\item{footnote_arg}{list, footnotes info}
}
\value{
kableExtra object
}
\description{
Function to print PDF table
}
