% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_prices.R
\name{price_calc}
\alias{price_calc}
\title{Calcolo prezzi}
\usage{
price_calc(pec, marg = 2, digits = DIGITS.PRICE, dazi = 0, b2b = TRUE)
}
\arguments{
\item{pec}{numeric cost}

\item{marg}{numeric margin on cost}

\item{digits}{integer rounding}

\item{dazi}{numeric dazi}

\item{b2b}{logical if b2b prices = 1}
}
\value{
numeric vector
}
\description{
Calcolo prezzi
}
